atlas_backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
Pillow==9.5.0
drf-yasg==1.21.5
python-dotenv==1.0.0